requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.35.24
botocore==1.35.24
cffi==1.17.1
cryptography==43.0.1
Django==4.2.16
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
